# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for a freelancer/service-provider business management app
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /team:
    get:
      operationId: getTeam
      tags: [account]
      summary: Get the current user's active team and its members
      responses:
        "200":
          description: Team with members
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamWithMembers"
        "404":
          description: User has no team yet
...
  /team/create:
    post:
      operationId: createTeam
      tags: [account]
      summary: Create a new team and become its owner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamInput"
      responses:
        "201":
          description: Team created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamWithMembers"
  /team/join:
    post:
      operationId: joinTeam
      tags: [account]
      summary: Join an existing team using an invite code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamJoinInput"
      responses:
        "200":
          description: Joined team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamWithMembers"
        "400":
          description: Invite code not found
        "409":
          description: Already a member of this team
  /team/switch:
    post:
      operationId: switchTeam
      tags: [account]
      summary: Switch the active team
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamSwitchInput"
      responses:
        "200":
          description: Active team switched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
        "403":
          description: Not a member of this team
...
  /team/members/{userId}:
    patch:
      operationId: updateTeamMemberRole
      tags: [account]
      summary: Change a member's role in the active team (owner only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamMemberRoleInput"
      responses:
        "200":
          description: Updated team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamWithMembers"
        "403":
          description: Not allowed
        "404":
          description: Member not found
    delete:
      operationId: removeTeamMember
      tags: [account]
      summary: Remove a member from the active team (owner only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamWithMembers"
        "403":
          description: Not allowed
        "404":
          description: Member not found
...
components:
  schemas:
...
    Me:
      type: object
      required: [id, email, name, teamId, teamName, role, teams]
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        teamId:
          type: ["string", "null"]
        teamName:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
        teams:
          type: array
          items:
            $ref: "#/components/schemas/TeamSummary"
    TeamSummary:
      type: object
      required: [id, name, role]
      properties:
        id:
          type: string
        name:
          type: string
        role:
          type: string
    TeamSwitchInput:
      type: object
      required: [teamId]
      properties:
        teamId:
          type: string
          minLength: 1
    TeamMember:
      type: object
      required: [id, name, email, role]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
    TeamWithMembers:
      type: object
      required: [id, name, inviteCode, createdAt, members]
      properties:
        id:
          type: string
        name:
          type: string
        inviteCode:
          type: string
        createdAt:
          type: string
          format: date-time
        members:
          type: array
          items:
            $ref: "#/components/schemas/TeamMember"
    TeamInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
    TeamJoinInput:
      type: object
      required: [inviteCode]
      properties:
        inviteCode:
          type: string
          minLength: 1
...
    TeamMemberRoleInput:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [owner, member]
...