# lib/api-zod/src/generated/types/teamSwitchInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification for a freelancer/service-provider business management app
 * OpenAPI spec version: 0.1.0
 */

export interface TeamSwitchInput {
  /** @minLength 1 */
  teamId: string;
}